kernel/services/ndvi_engine/requirements.txt
# SAHOOL NDVI Engine Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
nats-py>=2.6.0
numpy>=1.26.0
python-dateutil>=2.8.2
